Miscellaneous little fixes (#14)
* Disable nano repair if configured off

* Fix faction selection on object creation

The misc_object listbox generates scripts for each button when the
dialog opens, so the faction effectively cannot be changed in the
dialog. The workaround is to inject the faction into the script when the
button is clicked.

A cleaner fix would be to refactor the listbox or create a control which
is a scrollable list of GuiButton, but that seems like a change with
bigger impact.

* Add logging from Lua scripts

* Remove reputation display from relay screens

* Remove clock display from the relay screens

* Docking changes from commit 55e799261 on daid/EmptyEpsilon

Minor changes due to differences in the math library and SeriousProton
version.

* Expose getDockingState() for scripts

* Bugfix for docking: Ensure that GMs obey the rules

* Bugfix: Tutorial controls are linked to ship

Player ship information is managed in `GameGlobalInfo::update()`.

When starting a tutorial, the screens and controls are created before
the player information is set in `update()` and are inoperable. Fix is
to set the player information forcibly before creating the screens and
controls.

* Tutorial script updates

Changed factions for ships ("Human Navy" to "UCN", "Kraylor" to
"corsair", "Independent" remains the same) in tutorials.

Weapons tutorial now checks front and rear shields for damage when
teaching shield usage rather than just rear shield, in case the CPU ship
is attacking from the front.

Minor updates scripts to reflect Lua API changes (e.g.,
`commandSetTarget` now requires a second argument).

* Remove unnecessary files